seperated app and server into seperate files to avoid issues with tests accidently keeping server open
